bitwindow: shuffle around tabs
