+++ clients/bitwindow/lib/pages/sidechains_page.dart
@@ -2,8 +2,8 @@ import 'dart:math';
 
 import 'package:auto_route/auto_route.dart';
 import 'package:bitwindow/gen/wallet/v1/wallet.pbgrpc.dart';
-import 'package:bitwindow/pages/send_page.dart';
 import 'package:bitwindow/pages/sidechain_activation_management_page.dart';
+import 'package:bitwindow/pages/wallet_page.dart';
 import 'package:bitwindow/providers/sidechain_provider.dart';
 import 'package:bitwindow/providers/transactions_provider.dart';
 import 'package:bitwindow/servers/api.dart';
