+++ clients/sail_ui/lib/widgets/chips/action_header_chip.dart
@@ -25,49 +25,3 @@ class ActionHeaderChip extends StatelessWidget {
     );
   }
 }
-
-class DialogHeaderChip extends StatelessWidget {
-  final DialogType dialogType;
-  final String action;
-
-  const DialogHeaderChip({
-    super.key,
-    required this.action,
-    required this.dialogType,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = SailTheme.of(context);
-
-    return DecoratedBox(
-      decoration: BoxDecoration(
-        borderRadius: const BorderRadius.all(Radius.circular(4)),
-        color: theme.colors.chip,
-      ),
-      child: Padding(
-        padding: const EdgeInsets.all(
-          SailStyleValues.padding04,
-        ),
-        child: SailRow(
-          spacing: SailStyleValues.padding08,
-          children: [
-            SailSVG.fromAsset(findIconForType(dialogType)),
-            SailText.secondary12(action, bold: true),
-          ],
-        ),
-      ),
-    );
-  }
-
-  SailSVGAsset findIconForType(DialogType dialogType) {
-    switch (dialogType) {
-      case DialogType.info:
-        return SailSVGAsset.iconInfo;
-      case DialogType.error:
-        return SailSVGAsset.iconFailed;
-      case DialogType.success:
-        return SailSVGAsset.iconSuccess;
-    }
-  }
-}
